Added feature: allow email address as user name for imap.
Some imap servers (for example, gmx.com) requires email as user name to login.
Sign `@' is a delimeter for network server name in folders-file, so use `*'
instead of it in user names and convert it to `@' in place. You have to write
your real user name `xxx@gmx.com' like `xxx*gmx.com' in ~/.folders or config
file.