Remove usage of kill file to adjust message counter.
The kill file contains cons' with ranges of UIDs that are considered
'killed', i.e. hidden from user's view.  But these ranges cannot be
used to adjust counter of total messages in a IMAP4 mailbox: A range
of (1 . 100) does not necessarily refer to 100 messages.  E.g. if
there are 100 messages in the mailbox with UID ranging from 1 to 100
and you delete the first message, the range (1 . 100) refers to 99
messages.