* elmo-imap4.el (elmo-folder-delete-messages-plugged): Use smarter mechanism to expunge messages.
If there are already messages in the mailbox with the \Deleted flag
set, use following algorithm to delete only our messages:

If `elmo-imap4-use-uid' is non-nil, and the server supports the
UIDPLUS extension, use the UID EXPUNGE command.  Otherwise temporarily
remove \Deleted flag, call EXPUNGE, and restore the flag.

If `elmo-imap4-use-uid' is nil, use old mechanism and prompt the user.