fix the use of nconc in elmo-get-passwd
Previous change to elmo-get-passwd replaced append with nconc, but the side
effect of nconc is only visible when the extended list is non-empty. Instead
of fixing by storing the nconc result with setq use plain cons to put new
password at the front (and not at the end, as it was before).

Putting new data at the front should make no difference to an alist: there
is no item with the same key (because if it is then it is returned earlier).
Also, putting at the front is marginally faster (because we do not have to
re-traverse the alist to get to its end).