improve performance of elmo-list-insert
When inserting into a list after a given element the list was traversed
three times (first with memq to find the given element, then again with memq
to find the cdr of the given element, and finally with nconc to merge the
modified list with its original tail). Now the list is traversed only once.

This improves performance of wl-summary-sort-by-subject by about 10%, at
least in my setup.

Note: when the given element is not in the list, there are still two
traversals, but this use-case seem to not affect the performance.