elmo-split calls external functions with arguments
elmo-split called a function as ACTION, but it was called with no
arguments.  It was unuseful on lexical binding.  Now the function
is called with two arguments, folder and message number.
Furthermore, if it returns non-nil value, do `delete' ACTION.

* elmo/elmo-split.el (elmo-split-rule): Update docstring.
(elmo-split-subr): Call function with args.  Substitute function's
return value for delete-substance variable.

* doc/wl-ja.texi (Split messages): Update item function.

* doc/wl.texi (Split messages): Ditto.