Fix wl-message-follow-current-entity fails
mail-reply-buffer is an ancient Emacs's varaible, which was
renamed to mail-reply-action in 1996.  It seems that Wanderlust
doesn't need to use the common variable name.  So rename it to
have wl- prefix.  And, prepare the varaible to fix the original
problem.

Cf. https://github.com/wanderlust/wanderlust/issues/211

* wl/wl-draft.el (wl-draft-reply-buffer): Renamed from
mail-reply-buffer.
(wl-draft-add-in-reply-to, wl-draft-yank-from-mail-reply-buffer)
(wl-draft-insert-current-message, wl-draft-insert-get-message)
(wl-draft-yank-to-draft-buffer, wl-draft-yank-original): Adjust
users.

* wl/wl-message.el (wl-message-follow-current-entity): Let-bind
wl-mail-reply-buffer for wl-draft-yank-from-mail-reply-buffer.

* wl/wl-summary.el (mail-reply-buffer): Remove definition.