Refine adding Mail-Followup-To: field feature
When `wl-insert-mail-followup-to' is t, Mail-Followup-To: field is
added at sending and previewing time.  When its value is `reply',
the field is inserted in creating draft buffer.  It is the
previous behavior.
Cf. https://github.com/wanderlust/wanderlust/issues/239

wl-draft-make-mail-followup-to function now uses user's address in
argument when available and doesn't strip full-name from user's
address when `wl-draft-reply-use-address-with-full-name' is
non-nil.

* wl/wl-draft.el (wl-draft-make-mail-followup-to): Don't use
`wl-from' if argument contains user's address.  Don't strip
full-name from `wl-from' when
`wl-draft-reply-use-address-with-full-name' is non-nil.
(wl-draft-insert-mail-followup-to): New function.
(wl-draft-reply): Insert Mail-Followup-To: field only when
`wl-insert-mail-followup-to' value is 'reply'.
(wl-draft-send-hook-setup): New function.
(wl-draft-send): Use wl-draft-send-hook-setup.

* wl/wl-mime.el (wl-draft-preview-message): Use wl-draft-send-hook-setup.

* wl/wl-vars.el (wl-insert-mail-followup-to): Update definition as
above.

* doc/wl.texi (wl-insert-mail-followup-to):
* doc/wl-ja.texi (wl-insert-mail-followup-to): Updated.